Separated pcs prover and verifier
