+++ prover/src/lib.rs
@@ -1,10 +1,10 @@
 //! Crate describing modules used by main
 
 /// TODO: Add documentation
-pub mod skyscraper_pow;
-/// TODO: Add documentation
 pub mod skyscraper;
 /// TODO: Add documentation
-pub mod skyscraper_for_whir;
-/// TODO: Add documentation
 pub mod utils;
+/// TODO: Add documentation
+pub mod whir_utils;
+/// TODO: Add documentation
+pub mod sumcheck_utils;
